select which routine to run using command line arguments

+++ bin/fetch-xml-products-and-process.sh
@@ -16,40 +16,51 @@ F_LAST_SUB=$ZUNKAPATH/xml/aldo/aldo-products-substitution.xml
 # Log file.
 F_LOG=$ZUNKAPATH/log/aldo/aldo-xml-products.log
 
+# If today file not downloaded or have a invalid size.
+EXIST_TODAY_XML_FILE=$(ls -lh $ZUNKAPATH/xml/aldo/ | grep -E aldo-products-$(date +%Y-%h-%d)-[0-9]{6}\.xml | cut -d " " -f5 | sed "s/[A-Z,a-z]//" | awk '$1>100')
+if [[ -z $EXIST_TODAY_XML_FILE ]]; then
+    # Get xml file.
+    echo Downloading XML file.
+    # RESULT=`curl "http://webservice.aldo.com.br/asp.net/ferramentas/integracao.ashx?u=146612&p=zunk4c" \
+    # RESULT=`curl "https://www.zunka.com.br/xml/zoom/zoom-produtos.xml" \
+    RESULT=`curl "http://webservice.aldo.com.br/asp.net/ferramentas/integracao.ashx?u=146612&p=zunk4c" \
+        -s -w "%{time_total} %{size_download}" \
+        -o $F_BACKUP` 
 
-# Get xml file.
-echo Downloading XML file.
-# RESULT=`curl "http://webservice.aldo.com.br/asp.net/ferramentas/integracao.ashx?u=146612&p=zunk4c" \
-# RESULT=`curl "https://www.zunka.com.br/xml/zoom/zoom-produtos.xml" \
-RESULT=`curl "http://webservice.aldo.com.br/asp.net/ferramentas/integracao.ashx?u=146612&p=zunk4c" \
-    -s -w "%{time_total} %{size_download}" \
-    -o $F_BACKUP` 
+    TIME=`echo $RESULT | cut -d " " -f1`
+    SIZE=`echo $RESULT | cut -d " " -f2`
+    # Kilobytes.
+    SIZE=`expr $SIZE / 1024`
+    # Megabytes.
+    # SIZE=`expr $SIZE / 1048576`
 
-TIME=`echo $RESULT | cut -d " " -f1`
-SIZE=`echo $RESULT | cut -d " " -f2`
-# Kilobytes.
-SIZE=`expr $SIZE / 1024`
-# Megabytes.
-# SIZE=`expr $SIZE / 1048576`
+    # Log time and size.
+    printf "$(date +%FT%T%:z)  %.1fs  %.0fkb  $F_BACKUP\n" $TIME $SIZE >> $F_LOG
 
-# Log time and size.
-printf "$(date +%FT%T%:z)  %.1fs  %.0fkb  $F_BACKUP\n" $TIME $SIZE >> $F_LOG
+    # Copy as last xml file.
+    cp $F_BACKUP $F_LAST
 
-# Copy as last xml file.
-cp $F_BACKUP $F_LAST
+    # Substitution file.
+    F_SUB=$GS/aldowsc/bin/substitution-list.txt
 
-# Substitution file.
-F_SUB=$GS/aldowsc/bin/substitution-list.txt
+    # Create XML with Substitutions.
+    echo Creating xml with substitutions...
+    sed -f $F_SUB $F_LAST > $F_LAST_SUB
 
-# Create XML with Substitutions.
-echo Creating xml with substitutions...
-sed -f $F_SUB $F_LAST > $F_LAST_SUB
+    # Process xml file.
+    echo Processing XML file...
+    if [[ $RUN_MODE == production ]]; then
+        RUN_MODE=production aldowsc < $F_LAST_SUB
+    else
+        go build
+        ./aldowsc < $F_LAST_SUB
+    fi
+fi
 
-# Process xml file.
-echo Processing XML file...
+# Update stock quantity.
 if [[ $RUN_MODE == production ]]; then
-    RUN_MODE=production aldowsc < $F_LAST_SUB
+    RUN_MODE=production aldowsc --update-stock-quantity 
 else
-    go build
-    ./aldowsc < $F_LAST_SUB
+    go build $GS/aldowsc
+    $GS/aldowsc/aldowsc --update-stock-quantity
 fi
